Modificando os métodos createTCC e deleteTCC porém o delete esta incompleto

+++ gtcc/src/main/java/br/gtcc/gtcc/services/impl/neo4j/TccServices.java
@@ -1,9 +1,5 @@
 package br.gtcc.gtcc.services.impl.neo4j;
 
-import br.gtcc.gtcc.services.spec.TccInterface;
-import br.gtcc.gtcc.model.neo4j.Tcc;
-import br.gtcc.gtcc.model.neo4j.repository.TccRepository;
-
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,9 +23,45 @@ public class TccServices implements TccInterface<Tcc, String> {
     @SuppressWarnings("unlikely-arg-type")
     @Override
     public Tcc createTcc(Tcc tcc) {
-        if (tcc != null && tcc.getId() == null) {
-            return tccRepository.save(tcc);
-        } 
+       
+        System.out.println( "\n Tcc na service I " + tcc.toString());
+
+        String idAluno = tcc.getIdAluno();
+
+        String idOrientador = tcc.getIdOrientador();
+
+        if((idAluno != " " || idAluno != null) && (idOrientador != " " || idOrientador != null) ){
+
+            
+            Boolean existsAluno = this.usersRepository.existsById(idAluno);
+            Boolean existsOrientador = this.usersRepository.existsById(idOrientador); 
+            
+
+            if ((tcc.getId() != null || tcc.getId() != " ") && (existsAluno == true && existsOrientador == true)) {
+                
+                System.out.println( "\nTcc na service III" + tcc.toString());
+                
+                Users aluno = this.usersRepository.findById(idAluno).get();
+                Users orientador = this.usersRepository.findById(idOrientador).get();
+
+                tcc.setAluno(aluno);
+                tcc.setOrientador(orientador);
+                
+                if(orientador.getUserType().equals("COORDENADOR") || orientador.getUserType().equals("PROFESSOR")){}
+                orientador.getTccsGerenciados().add(tcc);
+                usersRepository.save(orientador);
+
+                return tccRepository.save(tcc);
+
+            }
+
+            if( (this.tccRepository.existsById(tcc.getId())) && (existsAluno == true && existsOrientador == true)){
+            
+                System.out.println( "\n Tcc na service IV" + tcc.toString());
+                return tccRepository.save(tcc);
+            }
+        }
+ 
         return null;
     }
 
@@ -56,20 +88,33 @@ public class TccServices implements TccInterface<Tcc, String> {
 
     @Override
     public Tcc deleteTCC(String id) {
-        Tcc delTcc = this.getTCC(id);
-        if (delTcc != null) {
-            tccRepository.delete(delTcc);
-        } else {
-            throw new IllegalArgumentException("O Tcc fornecido é inválido ou já possui um ID.");
+        
+        if( id != " " || id != null || id != "")  {
+
+            Tcc delTcc = this.getTCC(id);
+            if (delTcc != null) {
+                tccRepository.delete(delTcc);
+            } else {
+                throw new IllegalArgumentException("O Tcc fornecido é inválido ou já possui um ID.");
+            }
         }
+        
         return null;
     }
 
     @Override
     public List<Tcc> getAllTCC() {
-        return tccRepository.findAll();
+        
+        if(this.tccRepository.count() > 0){
+
+            return tccRepository.findAll();
+    
+        }
+
+        return null;
     }
 
+    @SuppressWarnings("unused")
     @Override
     public Tcc getTCC(String id) {
         if(id != null || id != " "){
